According to the FilterFactory manual, all operations are on signed 32-bit integers
git-svn-id: https://svn.viathinksoft.com/svn/filter_foundry/trunk@299 1b5a0048-3058-4200-831f-25aa9e3a14b9

+++ symtab.h
@@ -29,7 +29,8 @@
 #ifdef FP_VALUE
 typedef double value_type;
 #else
-typedef long value_type;
+// According to the FilterFactory manual, all operations are on signed 32-bit integers
+typedef int value_type;
 #endif
 typedef value_type (*pfunc_type)(value_type,...);
